% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_and_validity.R
\name{analysis_validity}
\alias{analysis_validity}
\title{批量分析效度}
\usage{
analysis_validity(
  data,
  subject_col,
  method_col,
  criterion_method,
  test_method,
  value_cols,
  trial_col = NULL,
  method = "mean",
  swc_threshold = NULL,
  var_names = NULL
)
}
\arguments{
\item{criterion_method}{一个字符串，指定 `method_col` 中代表**金标准**的水平名称。}

\item{test_method}{一个字符串，指定 `method_col` 中代表**测试方法**的水平名称。}
}
\description{
(推荐) 用于效度分析，通过明确指定金标准和测试方法，进行全面的比较...
}
\examples{
# 使用内置的 validity_data 数据集进行效度分析
analysis_validity(
  data = validity_data,
  subject_col = "Subject",
  method_col = "Device",
  criterion_method = "GasAnalyzer",
  test_method = "PortableDevice",
  value_cols = "VO2max"
)
}
