% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{descriptive_data}
\alias{descriptive_data}
\title{描述性统计示例数据集}
\format{
一个包含 60 行和 4 列的数据框:
\describe{
  \item{Subject}{运动员ID}
  \item{Session}{测试日期, 包含 "Day1" 和 "Day2" 两个水平}
  \item{Trial}{试次编号 (1, 2, 3)}
  \item{JumpHeight}{垂直纵跳高度 (cm)}
}
}
\usage{
descriptive_data
}
\description{
模拟10名运动员在2个不同日期，分别进行3次垂直纵跳测试的数据。
}
\keyword{datasets}
