% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reliability_data}
\alias{reliability_data}
\title{信度分析示例数据集}
\format{
一个包含 90 行和 4 列的数据框:
\describe{
  \item{Subject}{参与者ID}
  \item{Session}{测试时间点, 包含 "Week1", "Week2", "Week4" 三个水平}
  \item{Trial}{试次编号 (1, 2)}
  \item{Score}{力量测试得分 (N)}
}
}
\usage{
reliability_data
}
\description{
模拟15名参与者在3个不同时间点 (Week1, Week2, Week4)，
分别进行2次力量测试的数据。
}
\keyword{datasets}
