% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validity_data}
\alias{validity_data}
\title{效度分析示例数据集}
\format{
一个包含 40 行和 4 列的数据框:
\describe{
  \item{Subject}{个体ID}
  \item{Device}{测量设备名称, "GasAnalyzer" (金标准) 和 "PortableDevice" (测试设备)}
  \item{Trial}{试次编号 (此数据集中恒为1)}
  \item{VO2max}{最大摄氧量 (ml/kg/min)}
}
}
\usage{
validity_data
}
\description{
模拟20名个体分别使用金标准设备 (气体分析仪) 和一个便携式新设备
测量最大摄氧量 (VO2max) 的数据。每个测量被视为单次试次。
}
\keyword{datasets}
